+++ src/server/database/schema/business/models/payments-settings.ts
@@ -1,4 +1,4 @@
-import { integer, numeric, text } from 'drizzle-orm/pg-core'
+import { numeric, text } from 'drizzle-orm/pg-core'
 
 import { payment_currency_type } from '~/server/database/schema/business/enums/payment-currency-type'
 
@@ -9,7 +9,7 @@ export const payments_settings = business.table('payments_settings', {
   ...id,
 
   currency: payment_currency_type('currency').notNull(),
-  extraTaxRate: integer('extra_tax_rate').notNull(),
+  extraTaxRate: numeric('extra_tax_rate').notNull(),
   label: text('label').notNull(),
   tvaRate: numeric('tva_rate').notNull(),
